requirements.txt
apyori==1.1.1
amqp==2.3.2
APScheduler==3.5.3
bs4==0.0.1
certifi==2019.3.9
chardet==3.0.4
cloudpickle==0.6.0
cx-Oracle==6.2.1
cycler==0.10.0
dj-static==0.0.6
Django==1.11
django-cors-headers==2.4.0
djangorestframework==3.8.2
djongo==1.2.33
eli5==0.10.1
gevent==1.4.0
greenlet==0.4.15
idna==2.7
Jinja2==2.10.1
keras==2.1.5
kiwisolver==1.1.0
numpy==1.15.2
pandas==0.25.1
psycopg2-binary==2.8.3
python-dateutil==2.7.3
pyttsx3==2.7
pytz==2018.5
requests==2.19.1
scikit-learn==0.21.3 
scipy==1.1.0
seaborn==0.9.0
selenium==3.141.0
six==1.11.0
static3==0.7.0
tzlocal==1.5.1
urllib3==1.23
vine==1.1.4
xlrd==1.2.0

